+++ eelbrain/table.py
@@ -1,4 +1,4 @@
 """Create tables from data-objects"""
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
 
-from ._table import difference, frequencies, melt, stats, repmeas
+from ._table import difference, frequencies, melt, melt_ndvar, stats, repmeas