+++ doc/source/changes.rst
@@ -12,6 +12,7 @@ New in 0.16
   current definition.
 * :meth:`MneExperiment.make_mov_ttest` and
   :meth:`MneExperiment.make_mov_ga_dspm` are fixed but require PySurfer 0.6.
+* New function: :func:`table.melt_ndvar`.
 * API: :mod:`plot.brain` function signatures changed slightly to accommodate
   more layout-related arguments.